Update doctest for NumPy 2.0.1

+++ gymnasium/wrappers/transform_reward.py
@@ -77,7 +77,7 @@ class ClipReward(TransformReward[ObsType, ActType], gym.utils.RecordConstructorA
         >>> _ = env.reset()
         >>> _, rew, _, _, _ = env.step(1)
         >>> rew
-        np.float64(0.5)
+        0.5
 
     Change logs:
      * v1.0.0 - Initially added
